mypy fix

+++ torchgeo/datasets/vhr10.py
@@ -289,7 +289,8 @@ class VHR10(NonGeoDataset):
         )
         with Image.open(filename) as img:
             array: "np.typing.NDArray[np.int_]" = np.array(img)
-            tensor = torch.from_numpy(array).float()
+            tensor = torch.from_numpy(array)
+            tensor = tensor.float()
             # Convert from HxWxC to CxHxW
             tensor = tensor.permute((2, 0, 1))
             return tensor
